refactor(comment_list): move out CommentList widget
